% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextableFuns.R
\name{cat_ft}
\alias{cat_ft}
\title{Create and format flextable}
\usage{
cat_ft(df, Digits = 3, Caption = NULL, Tag = NULL,
  noHeader = FALSE, theme_ = "theme_vanilla",
  first.row.names = FALSE, row.names.null = TRUE,
  parse.pattern = "_", parse.replace = " ", font_ = "Cambria",
  fontsizeH = 8, fontsizeB = 8, rotationH = NULL, rotationB = NULL,
  alignB = "left", formatter = NULL)
}
\arguments{
\item{df}{data frame for table}

\item{Digits}{number of decimal places for rounding}

\item{Caption}{table caption}

\item{Tag}{tag for pandoc-crossref}

\item{noHeader}{logical to use header}

\item{theme_}{flextable theme to apply}

\item{first.row.names}{use first row as column headers}

\item{row.names.null}{use row names}

\item{parse.pattern}{pattern to replace}

\item{parse.replace}{string to replace parse.pattern}

\item{font_}{font to use}

\item{fontsizeH}{header font size in pts}

\item{fontsizeB}{body font size in pts}

\item{rotationH}{logical vector of header columns to rotate}

\item{rotationB}{logical vector of body columns to rotate}

\item{alignB}{body cell text alignment}

\item{formatter}{function for formatting cells}
}
\description{
Similar to cat_ptr but uses flextable for word documents
}
