% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panderFuns.R
\name{cat_pir}
\alias{cat_pir}
\title{Format and add image using pander::pandoc.image.return}
\usage{
cat_pir(plot_g, Caption = NULL, Tag = NULL)
}
\arguments{
\item{plot_g}{plot to be added to document}

\item{Caption}{caption for plot}

\item{Tag}{tag to be referenced using pandoc-crossref}
}
\description{
Submitted plot is added to document with tags
}
