% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panderFuns.R
\name{cat_ptr}
\alias{cat_ptr}
\title{Create pandoc table}
\usage{
cat_ptr(
  df,
  Digits = 3,
  Caption = NULL,
  Justify = "left",
  Style = "simple",
  Split.table = Inf,
  Split.cells = Inf,
  Keep.line.breaks = TRUE,
  Tag = NULL,
  font.s = "\\\\footnotesize\\n",
  first.row.names = FALSE,
  row.names.null = TRUE,
  parse.pattern = "_",
  parse.replace = " ",
  Latex = TRUE
)
}
\arguments{
\item{df}{data frame for table}

\item{Digits}{number of decimal places for rounding}

\item{Caption}{table caption}

\item{Justify}{text justification}

\item{Style}{markdown table style}

\item{Split.table}{number of rows to split table across pages}

\item{Split.cells}{number of cells to splist table}

\item{Keep.line.breaks}{keep line breaks in cells}

\item{Tag}{tag for pandoc-crossref}

\item{font.s}{latex font size}

\item{first.row.names}{use first row as column headers}

\item{row.names.null}{use row names}

\item{parse.pattern}{pattern to replace}

\item{parse.replace}{string to replace parse.pattern}

\item{Latex}{is document latex}
}
\description{
Create pandoc table
}
