% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineFactors.R
\name{cFactor}
\alias{cFactor}
\title{Combine factors with different levels}
\usage{
cFactor(factors, ordered.levels = NULL)
}
\description{
while retaining original levels of originals. submit factors as dataframe or list. ordered.levels not implemented but maybe pass the desired order of output levels
}
