% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilutionCalc.R
\name{conv_multiunit_}
\alias{conv_multiunit_}
\title{Modified conv_multiunit from measurements package}
\usage{
conv_multiunit_(x = 1, from, to)
}
\arguments{
\item{x}{numeric vector of measurement value}

\item{from}{initial units}

\item{to}{output units}
}
\description{
Allows from and to units to have no whitespace between units
}
