% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFuns.R
\name{dbConnect_}
\alias{dbConnect_}
\title{Wrapper to connect to db}
\usage{
dbConnect_(Group = "EPA_harshadb",
  defaultFile = file.path(gfuns::sg("ws"), ".my.cnf"), verbose = FALSE,
  ...)
}
\arguments{
\item{Group}{Group identifier found in .my.cnf}

\item{defaultFile}{file where Group credentials are found}

\item{verbose}{provide feedback to user}
}
\description{
Connect to database using RMariaDB. May want to add secure credentials method.
}
