% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilutionCalc.R
\name{dilCalc}
\alias{dilCalc}
\title{Dilution calculations}
\usage{
dilCalc(x)
}
\arguments{
\item{x}{data frame with columns v = c("c1", "v1", "c2", "v2"), value = value for variables with NA for variable to solve, units = units of values}
}
\description{
Quick calculation of concentration and volume of a dilution. Note it does no checking to make sure units are consistent. Three of four parameters must be supplied
}
