% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirCompare.R
\name{dirCompare}
\alias{dirCompare}
\title{Recursively compare two directories}
\usage{
dirCompare(dir1, dir2, Recursive = FALSE, dir1.excludes = NULL,
  compAttributes = FALSE)
}
\arguments{
\item{dir1}{directory to compare}

\item{dir2}{directory to compare}

\item{Recursive}{logical to perform recursive comparison}

\item{dir1.excludes}{character vector of files to exclude}

\item{compareAttributes}{logical to compare file attributes}
}
\description{
Recursively compare two directories
}
