% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_replacement.r
\name{exposurePR}
\alias{exposurePR}
\title{Wrapper to partialReplacement}
\usage{
exposurePR(
  flowTime = NULL,
  flowRate = NULL,
  volume = NULL,
  fraction = NULL,
  flowTimeUnits = "min",
  flowVolumeUnits = "l",
  totalTime,
  totalTimeUnits = "hr",
  concentration = NULL,
  concentrationUnits = "mg_per_l",
  contaminant = NA
)
}
\arguments{
\item{flowTime}{Time of flow}

\item{flowRate}{Rate of flow}

\item{volume}{Volume of reservoir}

\item{fraction}{Fraction replacement}

\item{totalTime}{total duration of exposure}

\item{totalTimeUnits}{total duration of exposure units}

\item{concentration}{exposure concentration}

\item{concentrationUnits}{exposure concentration units}

\item{contaminant}{name of contaminant}

\item{timeUnits}{= "min" use units as defined in package measurements}

\item{volumeUnits}{= "l" use units as defined in package measurements}
}
\description{
Generates partial replacement based on experimental conditions
}
