% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFuns.R
\name{get_PrimaryKeys}
\alias{get_PrimaryKeys}
\title{Get primary keys from MySQL table}
\usage{
get_PrimaryKeys(tName, Group = "EPA_harshadb")
}
\arguments{
\item{tName}{table name for which requesting primary keys}

\item{Group}{Group name linked to my.cnf with connection information}
}
\value{
Returns a one column (COLUMN_NAME) data.table with primary key fields.
}
\description{
Function will request table information from MySQL database information_schema.table_constraints table.
}
