% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepNegate.R
\name{greplN}
\alias{greplN}
\title{Function to apply negative regex in grepl}
\usage{
greplN(neg, string, ...)
}
\arguments{
\item{neg}{character string to be negated}

\item{string}{character string to search}

\item{...}{parameters to be passed to grep}
}
\description{
Function to apply negative regex in grepl
}
