% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepNegate.R
\name{grepN}
\alias{grepN}
\title{Function to apply negative regex in grep}
\usage{
grepN(neg, string, value = TRUE, ...)
}
\arguments{
\item{neg}{character string to be negated}

\item{string}{character string to search}

\item{value}{passed to grep. default is true}

\item{...}{parameters to be passed to grep}
}
\description{
Function to apply negative regex in grep
}
