% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Arrange multiple ggplots with single legend}
\usage{
grid_arrange_shared_legend(
  ...,
  plots = NULL,
  colN = NULL,
  rowN = NULL,
  layOut = NULL,
  position = c("bottom", "right"),
  Title = NULL,
  TitleJust = 1,
  TitleSize = 18
)
}
\arguments{
\item{...}{plots}

\item{plots}{list of plots}

\item{layOut}{matrix for layout of plots}

\item{position}{where should the legend be}

\item{Title}{optional title}

\item{TitleJust}{Title justification 0 = right, 0.5 = center, 1 = left}

\item{TitleSize}{Title size in points}

\item{ncolN}{number of columns for plots grob}

\item{nrowN}{number of rows for plots grob}
}
\description{
Take a list of plots, arrange in a grid with a title
}
