% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFuns.R
\name{HHsummary}
\alias{HHsummary}
\title{Data Summary Stats for HarshaHabs data}
\usage{
HHsummary(x, confInt = 0.05, sigFig = NULL, ...)
}
\arguments{
\item{x}{numeric vector}

\item{confInt}{numeric value giving p value for confidence intervals. If NULL, none are calculated.}

\item{sigFig}{number of significant digits to return, if NULL, it is calculated based on standard deviation}

\item{...}{parameters to be passed to summary functions such as na.rm or na.omit}
}
\description{
Function used to consistently return summary data for inclusion in HarshaHabs databse.
}
\details{
Returns N, Min, Max, Lower Quartile, Median, Upper Quartile, Max, Mean, Standard Deviation, Standard Error of the Mean, Coefficient of Variation, and mean confidence intervals.
}
