% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFuns.R
\name{insert_sql3}
\alias{insert_sql3}
\title{Function to insert values into db table}
\usage{
insert_sql3(
  dataF,
  tName,
  replaceifexists = FALSE,
  Group = "EPA_harshadb",
  defaultFile = file.path(gfuns::sg("ws"), ".my.cnf"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{dataF}{data frame to be appended}

\item{tName}{table to appand data to}

\item{Group}{Group identifier found in .my.cnf}

\item{defaultFile}{file where Group credentials are found}

\item{verbose}{provide feedback to user}
}
\description{
Uses parameterized statement with option to replace if exists
}
\details{
replaceifexists = TRUE is slower
Note: REPLACE makes sense only if a table has a PRIMARY KEY or UNIQUE index. Otherwise, it becomes equivalent to INSERT, because there is no index to be used to determine whether a new row duplicates another.
}
