% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{LayoutHandler}
\alias{LayoutHandler}
\title{Create layouts for grid allowing for stacking of plots}
\usage{
LayoutHandler(n, col_ = 6, stack = 2)
}
\arguments{
\item{n}{number of total plots}

\item{stack}{number of plots which should be stacked}

\item{ncol_}{max number of columns}
}
\description{
Create layouts for grid allowing for stacking of plots
}
