% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRdata.R
\name{loadRData}
\alias{loadRData}
\title{Load subset or all .RData objects from submitted path}
\usage{
loadRData(Path = gfuns::sg("lh_Robjects"), objectList = NULL)
}
\arguments{
\item{Path}{path to Robjects}

\item{objectList}{list of objects to load}
}
\description{
Load subset or all .RData objects from submitted path
}
