% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossrefFuns.R
\name{multiFig}
\alias{multiFig}
\title{Concatenate multiple figure references}
\usage{
multiFig(figpre = NULL, figpost, sep = ".")
}
\arguments{
\item{figpre}{pre tag string}

\item{figpost}{post tag string}

\item{sep}{paste separator}
}
\description{
Uses pandoc-crossref style
}
