% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossrefFuns.R
\name{multiRef}
\alias{multiRef}
\title{Generalized function to insert multiple figure or table references}
\usage{
multiRef(pre = NULL, post = NULL, ref = NULL, s = "ref",
  sep = ".")
}
\arguments{
\item{pre}{pre tag string}

\item{post}{post tag string}

\item{ref}{reference type - fig, tbl, etc}

\item{s}{ref or tag}

\item{sep}{separator for paste}
}
\description{
Concatenates multiple figure references for pandoc-crossref based on multiFig and multiTbl functions.
}
