% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plotHandler}
\alias{plotHandler}
\title{Plot handler}
\usage{
plotHandler(
  visPlot = FALSE,
  makeObject = FALSE,
  savePlots = TRUE,
  filename = NULL,
  path = NULL,
  plt = ggplot2::last_plot(),
  device = c("pdf", "png", "wmf"),
  dpi = 300
)
}
\arguments{
\item{visPlot}{logical to display plot}

\item{makeObject}{logical to save plots in environment}

\item{savePlots}{logical to save plots as image}

\item{filename}{filename to pass to ggsave_f}

\item{path}{path where plots should be saved to pass to ggsave_f}

\item{plt}{plot to be printed to pass to ggsave_f}

\item{device}{list of formats to be saved to pass to ggsave_f}

\item{dpi}{image resolution to pass to ggsave_f}
}
\description{
Function to save and display plots
}
