% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plotKknit}
\alias{plotKknit}
\title{Plot knit Function}
\usage{
plotKknit(chunkLabel, capt, plt, pos = "H", ...)
}
\arguments{
\item{chunkLabel}{text for chunk label which is also used for figure file name}

\item{capt}{text for caption}

\item{plt}{plot object to be placed}

\item{pos}{latex option for keeping the plot here}
}
\description{
This function dynamically produces a knitr chunk with appropriate captions and labels which is then knitted into the main document with cat.  Any chunk option could be added for full control over chunk.
}
