% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFuns.R
\name{pooled_sem}
\alias{pooled_sem}
\title{Data Summary Stats for Pooled Groups}
\usage{
pooled_sem(x, confInt = 0.05, ...)
}
\arguments{
\item{x}{numeric vector}

\item{confInt}{numeric value giving p value for confidence intervals. If NULL, none are calculated.}

\item{...}{parameters to be passed to summary functions such as na.rm or na.omit}
}
\description{
Function used by summaryBy or summarise to compute summary stats on pooled data.
}
\details{
Returns N, Min, Max, Lower Quartile, Median, Upper Quartile, Max, Mean, Standard Deviation, Standard Error of the Mean, Coefficient of Variation, and mean confidence intervals.
}
