% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFuns.R
\name{RPD}
\alias{RPD}
\title{Calculate Relative Percent Difference}
\usage{
RPD(x, sigFig = NULL)
}
\arguments{
\item{x}{vector of length 2}

\item{sigFig}{if not null, the number of significant digits to report}
}
\description{
Calculate Relative Percent Difference
}
