% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFuns.R
\name{send_sql3}
\alias{send_sql3}
\title{Function to send sql statement}
\usage{
send_sql3(sql, Group = "EPA_harshadb",
  defaultFile = file.path(gfuns::sg("ws"), ".my.cnf"), verbose = FALSE,
  CLIENT_MULTI_STATEMENTS_ = TRUE, ...)
}
\arguments{
\item{sql}{sql statement}

\item{Group}{Group identifier found in .my.cnf}

\item{defaultFile}{file where Group credentials are found}

\item{verbose}{provide feedback to user}

\item{CLIENT_MULTI_STATEMENTS_}{logical to use CLIENT_MULTI_STATEMENTS}
}
\description{
Submit sql statement using RMariaDB. May want to add secure credentials method.
}
