% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{separate_multi}
\alias{separate_multi}
\title{#' Function to modify ggplot facet_grid panel labels
#' @export
plot_labeller <- function(variable,value){
  return(plot_names[value])
}
Separate multiple columns}
\usage{
separate_multi(.data, .cols, .suffix, .sep)
}
\arguments{
\item{.data}{data frame with columns to split}

\item{.cols}{columns to split}

\item{.suffix}{suffix to add to each column name for expansion}

\item{.sep}{separator string in .cols}
}
\description{
Takes a data frame with multiple columns to split and returns resulting frame
}
