% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setpaths.R
\name{setWorkspace}
\alias{setWorkspace}
\title{Get workspace}
\usage{
setWorkspace()
}
\value{
character value representing workspace directory.
}
\description{
Helper for setpaths to establish workspace programatically
}
\details{
This function sets the workspace based on OS type and OneDrive availability.
}
