% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setpaths.R
\name{sg}
\alias{sg}
\title{Shortcut to Sys.getenv}
\usage{
sg(x)
}
\arguments{
\item{x}{Environment variable string}
}
\description{
Sys.getenv is too much typing
}
\details{
This function is included for backward compatibility. Paths are now in a list named path and are accessed via path$pathName or path[[`pathName`]]. Previously, paths were set in the system enironment. It is still capable of returning a system environment variable as it looks for x in the system environment before path.
}
