% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sigFunc}
\alias{sigFunc}
\title{Signature page table function for latex signature page}
\usage{
sigFunc(sigs, online = NULL, vspace = "0.75in")
}
\arguments{
\item{sigs}{list of names and affiliation to be included in signatures}

\item{online}{should name be on line}

\item{vspace}{latex space between signature entries}
}
\description{
supply name and titles and output dataframe appropriate for pander::pandoc.table. Used in SOPs and QAPP.  sigs is list of sigs and title
}
\examples{
sigs <- list(list(name = 'Joel Allen',
                  title = 'Principle Investigator',
                  online = 'test'),
             list(name = 'Michael Elovitz',
                  title = c('Branch Chief',
                            'Water Quality Management Branch')))
}
