% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{time2DT}
\alias{time2DT}
\title{Function to adjust Time values as read by read_excel}
\usage{
time2DT(date.v, time.v)
}
\arguments{
\item{date.v}{POSIX date vector}

\item{time.v}{POSIC time vector}
}
\description{
input is a dataframe with two columns labeled Date and Time. Both need to be in the same timezone.
}
\section{Value}{

output is a DateTime vector
}

