% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTime.R
\name{toUTC}
\alias{toUTC}
\title{Convert to UTC}
\usage{
toUTC(DT, tz = "UTC")
}
\arguments{
\item{DT}{vector of DateTimes in POSIXct}

\item{tz}{Timezone to convert to, defaults to "UTC"}
}
\description{
Given POSIXct convert time to UTC
}
