% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileConvert.R
\name{u2c}
\alias{u2c}
\title{Convert utf8 hex to character}
\usage{
u2c(x)
}
\arguments{
\item{x}{hex or character vector of utf8 character codes}
}
\description{
Wrapper for rawToChar to place utf8 symbols in text
}
\examples{
# Character codes can be found at https://www.utf8-chartable.de/
plusminus <- "0xb1"
tilde <- "0x7e"
# Codes don't have to be quoted
atsign <- 0x40
degreeSymbol <- 0xb0
asterisk <- 0x2a
u2c(plusminus)
# Vector
u2c(c(0x48, 0x65, 0x6c, 0x6c, 0x6f))
}
