% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFuns.R
\name{upsert_sql3}
\alias{upsert_sql3}
\title{Function to insert or update on duplicate key}
\usage{
upsert_sql3(
  dataF,
  tName,
  key = NULL,
  Group = "EPA_harshadb",
  defaultFile = file.path(gfuns::sg("ws"), ".my.cnf"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{dataF}{data frame to be appended}

\item{tName}{table to appand data to}

\item{Group}{Group identifier found in .my.cnf}

\item{defaultFile}{file where Group credentials are found}

\item{verbose}{provide feedback to user}
}
\description{
Insert data using parameterized statement. If duplicate key exists perform Update. Records with extant keys will be updated those without will be inserted using RMariaDB.
}
\details{
Return value is number of rows affected
}
