% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFuns.R
\name{append_sql3}
\alias{append_sql3}
\title{Function to append dataframe to db table}
\usage{
append_sql3(
  dataF,
  tName,
  Group = "EPA_harshadb",
  defaultFile = file.path(gfuns::sg("ws"), ".my.cnf"),
  header_ = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{dataF}{data frame to be appended}

\item{tName}{table to appand data to}

\item{Group}{Group identifier found in .my.cnf}

\item{defaultFile}{file where Group credentials are found}

\item{header_}{logical whether first row is header}

\item{verbose}{provide feedback to user}
}
\description{
Submit data to be appended to database table using RMariaDB. Output 0 or 1 based on success if verbose = TRUE May want to add secure credentials method.
}
