% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFuns.R
\name{cv}
\alias{cv}
\title{Coefficient of variation}
\usage{
cv(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{parameters to be passed to sd and mean}
}
\description{
Calculate coefficient of variation of vector
}
