% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ggsave_f}
\alias{ggsave_f}
\title{Custom ggsave function}
\usage{
ggsave_f(filename, path = NULL, plt = last_plot(), formats = c("pdf",
  "png", "wmf"), dpi = 300)
}
\arguments{
\item{filename}{filename}

\item{path}{path where plots should be saved}

\item{plt}{plot to be printed}

\item{formats}{list of formats to be saved}

\item{dpi}{image resolution}
}
\description{
Allows for multiple file formats
}
