% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbFuns.R
\name{send_sql3}
\alias{send_sql3}
\title{Function to send sql statement}
\usage{
send_sql3(
  dataF,
  tName,
  returnClass = "data.table",
  sql = "SELECT * FROM ",
  Group = "EPA_harshadb",
  defaultFile = file.path(gfuns::sg("ws"), ".my.cnf"),
  verbose = FALSE,
  CLIENT_MULTI_STATEMENTS_ = TRUE,
  ...
)
}
\arguments{
\item{dataF}{If given, data frame of key values to be used in parameterized statement.}

\item{tName}{table name}

\item{returnClass}{class of return object. Should be one of data.table, tibble, or data.frame.}

\item{sql}{sql statement. defaults to select}

\item{Group}{Group identifier found in .my.cnf}

\item{defaultFile}{file where Group credentials are found}

\item{verbose}{provide feedback to user}

\item{CLIENT_MULTI_STATEMENTS_}{logical to use CLIENT_MULTI_STATEMENTS}
}
\description{
Submit sql statement using RMariaDB. May want to add secure credentials method.
}
