% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setpaths.R
\name{setpaths}
\alias{setpaths}
\title{Set paths}
\usage{
setpaths(SubD = c("defaults", "data", "Robjects", "figures", "plots"))
}
\arguments{
\item{SubD}{character vector of subdirectories to use to build paths}
}
\description{
Set paths relative to working directory
}
