% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{timeExcel_Posix}
\alias{timeExcel_Posix}
\title{Time values in Excel spreadsheets are imported as Posix}
\usage{
timeExcel_Posix(date_, time_, tZ = Tzone)
}
\arguments{
\item{date_}{POSIX date}

\item{time_}{POSIX time}

\item{tZ}{timezone}
}
\description{
Defaults to the origin date and UTC timezone of microsoft excel which is non-standard. This function extracts the hour, minute, second values and creates a lubridate hms object with specified timezone which can then be combined with _Date to generate _DateTime value Timezone Etc/GMT-6 is Eastern Standard Time
}
